requirements.txt
requests==2.31.0
pandas==2.1.3
openpyxl==3.1.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.108.0
playwright==1.40.0
schedule==1.2.0
openai==1.6.1
python-dotenv==1.0.0
SpeechRecognition==3.10.1
PyAudio==0.2.14
watchdog
reportlab
streamlit
langgraph
langchain
langchain-groq
twilio
gtts
pyttsx3